New Column created_by in Chamados table, Bug Fixes
Added a new column to store the creator of the chamado. Added the creator_by to the chamado page. Added the floating button from the chamado page to the Chamados page.  Removed the toastr css from the chamado scripts. Removed the value from the filter modal, so the browser doesnt send the warning.